requirements.txt
numpy>=1.19.2
pandas>=1.2.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
joblib>=1.0.0
pyyaml>=5.4.1